+++ javascript/node/selenium-webdriver/lib/webdriver.js
@@ -620,48 +620,54 @@ class WebDriver {
 
   /**
    * Schedules a command to wait for a condition to hold. The condition may be
-   * specified by a {@link webdriver.until.Condition}, as a custom function, or
-   * as a {@link webdriver.promise.Promise}.
+   * specified by a {@link until.Condition}, as a custom function, or as any
+   * promise-like thenable.
    *
-   * For a {@link webdriver.until.Condition} or function, the wait will repeatedly
+   * For a {@link until.Condition} or function, the wait will repeatedly
    * evaluate the condition until it returns a truthy value. If any errors occur
    * while evaluating the condition, they will be allowed to propagate. In the
-   * event a condition returns a {@link webdriver.promise.Promise promise}, the
-   * polling loop will wait for it to be resolved and use the resolved value for
-   * whether the condition has been satisified. Note the resolution time for
-   * a promise is factored into whether a wait has timed out.
+   * event a condition returns a {@link promise.Promise promise}, the polling
+   * loop will wait for it to be resolved and use the resolved value for whether
+   * the condition has been satisified. Note the resolution time for a promise
+   * is factored into whether a wait has timed out.
    *
-   * *Example:* waiting up to 10 seconds for an element to be present and visible
-   * on the page.
+   * Note, if the provided condition is a {@link until.WebElementCondition}, then
+   * the wait will return a {@link WebElementPromise} that will resolve to the
+   * element that satisified the condition.
+   *
+   * *Example:* waiting up to 10 seconds for an element to be present on the
+   * page.
    *
    *     var button = driver.wait(until.elementLocated(By.id('foo')), 10000);
    *     button.click();
    *
    * This function may also be used to block the command flow on the resolution
-   * of a {@link webdriver.promise.Promise promise}. When given a promise, the
-   * command will simply wait for its resolution before completing. A timeout may
-   * be provided to fail the command if the promise does not resolve before the
-   * timeout expires.
+   * of any thenable promise object. When given a promise, the command will
+   * simply wait for its resolution before completing. A timeout may be provided
+   * to fail the command if the promise does not resolve before the timeout
+   * expires.
    *
    * *Example:* Suppose you have a function, `startTestServer`, that returns a
-   * promise for when a server is ready for requests. You can block a `WebDriver`
+   * promise for when a server is ready for requests. You can block a WebDriver
    * client on this promise with:
    *
    *     var started = startTestServer();
    *     driver.wait(started, 5 * 1000, 'Server should start within 5 seconds');
    *     driver.get(getServerUrl());
    *
-   * @param {!(webdriver.promise.Promise<T>|
-   *           webdriver.until.Condition<T>|
-   *           function(!webdriver.WebDriver): T)} condition The condition to
+   * @param {!(promise.Promise<T>|
+   *           until.Condition<T>|
+   *           function(!WebDriver): T)} condition The condition to
    *     wait on, defined as a promise, condition object, or  a function to
    *     evaluate as a condition.
    * @param {number=} opt_timeout How long to wait for the condition to be true.
    * @param {string=} opt_message An optional message to use if the wait times
    *     out.
-   * @return {!webdriver.promise.Promise<T>} A promise that will be fulfilled
-   *     with the first truthy value returned by the condition function, or
-   *     rejected if the condition times out.
+   * @return {!(promise.Promise<T>|WebElementPromise)} A promise that will be
+   *     resolved with the first truthy value returned by the condition
+   *     function, or rejected if the condition times out. If the input
+   *     input condition is an instance of a {@link until.WebElementCondition},
+   *     the returned value will be a {@link WebElementPromise}.
    * @template T
    */
   wait(condition, opt_timeout, opt_message) {
@@ -679,12 +685,24 @@ class WebDriver {
     }
 
     var driver = this;
-    return this.flow_.wait(function() {
+    var result = this.flow_.wait(function() {
       if (promise.isGenerator(fn)) {
         return promise.consume(fn, null, [driver]);
       }
       return fn(driver);
     }, opt_timeout, message);
+
+    if (condition instanceof until.WebElementCondition) {
+      result = new WebElementPromise(this, result.then(function(value) {
+        if (!(value instanceof WebElement)) {
+          throw TypeError(
+              'WebElementCondition did not resolve to a WebElement: '
+                  + Object.prototype.toString.call(value));
+        }
+        return value;
+      }));
+    }
+    return result;
   }
 
   /**